Move Standfirst into Body (slices) section
