+++ common/views/components/Body/Body.config.json
@@ -1,0 +1,3 @@
+{
+  "name": "Body slices"
+}
