+++ common/views/components/Body/Standfirst/Standfirst.config.js
